+++ sha3/md_helper.c
@@ -129,7 +129,7 @@ SPH_XCAT(sph_, HASH)(void *cc, const void *data, size_t len)
 #endif
 {
 	SPH_XCAT(sph_, SPH_XCAT(HASH, _context)) *sc;
-	unsigned current;
+	size_t current;
 
 	sc = cc;
 #if SPH_64
@@ -138,7 +138,7 @@ SPH_XCAT(sph_, HASH)(void *cc, const void *data, size_t len)
 	current = (unsigned)sc->count_low & (SPH_BLEN - 1U);
 #endif
 	while (len > 0) {
-		unsigned clen;
+		size_t clen;
 #if !SPH_64
 		sph_u32 clow, clow2;
 #endif
